forbidden API: move rule about Collections.sort() to Lucene rules.
While this rule applies to core, it propose to replace calls to
Collections.sort() with calls to similar functions in Lucene. Thus it
only make sense for projects which already have Lucene on the classpath.

Similar rule about prefering FastStringReader was dropped, as
FastStringReader is being deprecated on the Lucene side.

Change-Id: I193a26e4f20f8c02394849e17f4e7ebe55e2e7ff